apply code cleanups: fix analyzer warnings and remove unused

+++ CBOR/PeterO/Cbor/CBORDoubleBits.cs
@@ -134,7 +134,7 @@ CBORUtilities.DoubleRetainsSameValueInSingle((long)obj);
         return origbits;
       }
       // Infinity and NaN
-      if (bits >= unchecked((long)(0x7ffL << 52)))
+      if (bits >= unchecked(0x7ffL << 52))
       {
         return origbits;
       }
